pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-agent-platform"
version = "0.1.0"
description = "基于 Python 的 AI Agent 平台"
authors = [{name = "AI Assistant"}]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "python-multipart==0.0.6",
    "sqlalchemy==2.0.25",
    "aiosqlite==0.19.0",
    "httpx==0.26.0",
    "tenacity==8.2.3",
    "chromadb==0.4.22",
    "pyyaml==6.0.1",
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.0",
    "loguru==0.7.2",
    "click==8.1.7",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
    "pytest-cov==4.1.0",
    "black==23.12.1",
    "ruff==0.1.11",
    "mypy==1.8.0",
]

[project.scripts]
agent-platform = "src.cli:main"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
